Tanks v0.7.5 - Changed the latency meter to show an average over the last second, fixed not being able to see other players moving in parties, and improved the item bar handling of bullet tiems. Tanks v0.7.4 - Rewrote the multiplayer system again, making it more secure by handling more things on the server side, added latency indicators when playing in parties, improved drawing of fancy bullet trails, and made other bug fixes and improvements

+++ src/main/java/tanks/gui/Firework.java
@@ -126,7 +126,7 @@ public class Firework extends Movable
 		else if (type == FireworkType.trail)
 		{	
 			Drawing.drawing.setColor(this.colorR, this.colorG, this.colorB, Math.max(0, Math.min(255, 255 - (int) (this.age * 255.0 / this.maxAge))));
-			double s = Math.max(0, this.size - (int) (this.age * this.size / this.maxAge));
+			double s = Math.max(0, this.size - (this.age * this.size / this.maxAge));
 			Drawing.drawing.fillOval(posX, posY, s, s);
 			
 			if (this.age >= this.maxAge)
